🔧 refactor: Rename app folder to redux
