+++ src/tests/index.test.js
@@ -4,7 +4,7 @@ import { cleanup } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import App from '../App';
-import store from '../app/store';
+import store from '../redux/store';
 
 afterEach(() => cleanup());
